Ensure colliding with a wall during dash will only exit dashing
It was previously possible for the delayed state exit call to exit other states,
such as ball possession, if the state transitioned to something beside dashing
in the one second default time delay. The desired behavior was to wait for 0.1
seconds, not 1, and only end the current state if it's still Dash.